myBank/requirements.txt
Flask
flask-wtf
flask-login
cassandra-driver
kafka-python
